innovation/app/requirements.txt
# ========================================
# CORE FRAMEWORK
# ========================================
fastapi==0.128.0
uvicorn==0.34.0
python-multipart==0.0.20
jinja2==3.1.6

# ========================================
# DATABASE (Otimizado para Vercel)
# ========================================
sqlalchemy==2.0.46
psycopg2-binary==2.9.10

# ========================================
# AUTHENTICATION & SECURITY
# ========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
slowapi==0.1.9

# ========================================
# VALIDATION & SETTINGS
# ========================================
pydantic==2.12.0
pydantic-settings==2.7.0
email-validator==2.2.0

# ========================================
# HTTP CLIENT
# ========================================
httpx==0.28.1

# ========================================
# AI - GOOGLE GEMINI (Essencial)
# ========================================
google-generativeai==0.8.3

# ========================================
# EMAIL - SENDGRID (Leve)
# ========================================
sendgrid==6.11.0

# ========================================
# UTILITIES
# ========================================
python-dotenv==1.0.1
